backend/requirements.txt
fastapi
uvicorn
pydantic
bycrypt
sqlalchemy
passliby